Respond to the case where an unitialized device is connected.
When a controller that hasn't been used before on a system power cycle is connected, it appears that a device may be (about 50% of the time) reported to the OuyaUnityActivity before it can be initialized by the OUYA SDK. It comes over with a playerNum of -1. The OuyaUnityBridge can now respond to this event by requesting another device check on the next frame. This fix seems to handle this condition.